[Fixed] Proper handling of UIListen for Installations (You broke crafting cRush =p) [Updated] Slight fix for resource ID's
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1490 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/structure/StructurePermissionList.h
@@ -84,15 +84,15 @@ public:
 
 	inline uint8 getPermission(const String& listname) {
 		if (listname == "BAN")
-			return pow(2, BAN);
+			return (uint8)pow(2, BAN);
 		else if (listname == "ENTRY")
-			return pow(2, ENTRY);
+			return (uint8)pow(2, ENTRY);
 		else if (listname == "HOPPER")
-			return pow(2, HOPPER);
+			return (uint8)pow(2, HOPPER);
 		else if (listname == "VENDOR")
-			return pow(2, VENDOR);
+			return (uint8)pow(2, VENDOR);
 
-		return pow(2, ADMIN);
+		return (uint8)pow(2, ADMIN);
 	}
 
 	inline bool isOwner(Player* player) {
